+++ changelog.md
@@ -6,4 +6,5 @@
   * _src --> used for creating database models and scriptella scripts (jade)
   * _workshop --> used primarily by Knodeo Workshop and infrequently modified by end users
 * Wired in `knodeo new --database {databasename}`
+* Wired in `knodeo run --migration` -- parameters not tested but should work
 * Tweaks to the gulp compilers (no impact to knodeo workshop users)